Link auth user to existing table
Added another column to the table to hold the uuid from an authorised user. Now I can grab the information that is linked to the user that is signed in instead of doing it manually. Still think that this can be done way better because at the moment a bunch of stuff has to be done in the backend.